filer bug fixed in repository, radius bug fixed in front, front changed(profiles, reviews)

+++ src/main/java/am/aca/bookingmanagement/dto/filter/FilteredPartnerResponseDetails.java
@@ -23,6 +23,6 @@ public class FilteredPartnerResponseDetails {
 
     private String imageUrl;
 
-    private Integer review;
+    private Double review;
 
 }
